python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
xgboost==2.0.3
redis==5.0.1
kafka-python==2.0.2
pydantic==2.5.3
python-multipart==0.0.6
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.1
